screen-translator/requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
pytesseract==0.3.10
PyQt6==6.6.1 